break out STM32F4 machine with new build tag
