+++ src/machine/machine_stm32.go
@@ -4,6 +4,19 @@ package machine
 
 // Peripheral abstraction layer for the stm32.
 
+const (
+	portA Pin = iota * 16
+	portB
+	portC
+	portD
+	portE
+	portF
+	portG
+	portH
+	portI
+	portJ
+)
+
 type PinMode uint8
 
 // Peripheral operations sequence:
