+++ targets/stm32f4disco.json
@@ -1,6 +1,6 @@
 {
   "inherits": ["cortex-m4"],
-  "build-tags": ["stm32f4disco", "stm32f407", "stm32"],
+  "build-tags": ["stm32f4disco", "stm32f407", "stm32f4", "stm32"],
   "cflags": [
     "-Qunused-arguments"
   ],
